pyproject.toml
[project]
name = "voice-input"
version = "0.1.0"
description = "Smart voice input with speaker identification, intent recognition, and text cleanup"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.0.0",
    "pyannote.audio>=3.1",
    "torch>=2.0",
    "torchaudio>=2.0",
    "sounddevice>=0.4",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voice_input"]
